test absolute paths added and folder renaming change
