+++ src/services/Api.ts
@@ -1,4 +1,4 @@
-import { forkJoin, Observable } from 'rxjs'
+import { forkJoin, Observable } from 'rxjs';
 import {
   ajax,
   AjaxRequest,
